BLE review part 2
